+++ tests/test_argument.py
@@ -12,15 +12,17 @@ class TestArgument(TestCase):
 
     def test_basic(self):
         a = Argument('| a = b ')
-        self.assertEqual(' a ', a.name)
-        self.assertEqual(' b ', a.value)
-        self.assertEqual(False, a.positional)
-        self.assertEqual(repr(a), "Argument('| a = b ')")
+        ae = self.assertEqual
+        ae(' a ', a.name)
+        ae(' b ', a.value)
+        ae(False, a.positional)
+        ae(repr(a), "Argument('| a = b ')")
 
     def test_anonymous_parameter(self):
+        ae = self.assertEqual
         a = Argument('| a ')
-        self.assertEqual('1', a.name)
-        self.assertEqual(' a ', a.value)
+        ae('1', a.name)
+        ae(' a ', a.value)
 
     def test_set_name(self):
         a = Argument('| a = b ')
@@ -28,32 +30,43 @@ class TestArgument(TestCase):
         self.assertEqual('| c = b ', a.string)
 
     def test_set_name_at_subspan_boundary(self):
+        ae = self.assertEqual
         a = Argument('|{{ a }}={{ b }}')
         a.name = ' c '
-        self.assertEqual('| c ={{ b }}', a.string)
-        self.assertEqual('{{ b }}', a.value)
+        ae('| c ={{ b }}', a.string)
+        ae('{{ b }}', a.value)
 
     def test_set_name_for_positional_args(self):
         a = Argument('| b ')
         a.name = a.name
         self.assertEqual('|1= b ', a.string)
 
-    def test_set_value(self):
+    def test_value_setter(self):
         a = Argument('| a = b ')
         a.value = ' c '
         self.assertEqual('| a = c ', a.string)
 
+    def test_value_deleter(self):
+        ae = self.assertEqual
+        a = Argument('| a = b ')
+        del a.value
+        ae('| a ', a.string)
+        del a.value
+        ae('|', a.string)
+
     def test_removing_last_arg_should_not_effect_the_others(self):
+        ae = self.assertEqual
         a, b, c = Template('{{t|1=v|v|1=v}}').arguments
         del c[:]
-        self.assertEqual('|1=v', a.string)
-        self.assertEqual('|v', b.string)
+        ae('|1=v', a.string)
+        ae('|v', b.string)
 
     def test_nowikied_arg(self):
+        ae = self.assertEqual
         a = Argument('|<nowiki>1=3</nowiki>')
-        self.assertEqual(True, a.positional)
-        self.assertEqual('1', a.name)
-        self.assertEqual('<nowiki>1=3</nowiki>', a.value)
+        ae(True, a.positional)
+        ae('1', a.name)
+        ae('<nowiki>1=3</nowiki>', a.value)
 
     def test_value_after_convertion_of_positional_to_keywordk(self):
         a = Argument("""|{{{a|{{{b}}}}}}""")
@@ -66,21 +79,23 @@ class TestArgument(TestCase):
             [a.name for a in parse('{{t|a|b|c}}').templates[0].arguments])
 
     def test_dont_confuse_subspan_equal_with_keyword_arg_equal(self):
+        ae = self.assertEqual
         p = parse('{{text| {{text|1=first}} | b }}')
         a0, a1 = p.templates[0].arguments
-        self.assertEqual(' {{text|1=first}} ', a0.value)
-        self.assertEqual('1', a0.name)
-        self.assertEqual(' b ', a1.value)
-        self.assertEqual('2', a1.name)
+        ae(' {{text|1=first}} ', a0.value)
+        ae('1', a0.name)
+        ae(' b ', a1.value)
+        ae('2', a1.name)
 
     def test_setting_positionality(self):
+        ae = self.assertEqual
         a = Argument("|1=v")
         a.positional = False
-        self.assertEqual('|1=v', a.string)
+        ae('|1=v', a.string)
         a.positional = True
-        self.assertEqual('|v', a.string)
+        ae('|v', a.string)
         a.positional = True
-        self.assertEqual('|v', a.string)
+        ae('|v', a.string)
         self.assertRaises(ValueError, setattr, a, 'positional', False)
 
     def test_parser_functions_at_the_end(self):
@@ -88,38 +103,36 @@ class TestArgument(TestCase):
         self.assertEqual(1, len(pfs))
 
     def test_section_not_keyword_arg(self):
+        ae = self.assertEqual
         a = Argument('|1=foo\n== section ==\nbar')
-        self.assertEqual(
-            (a.name, a.value), ('1', 'foo\n== section ==\nbar'))
+        ae((a.name, a.value), ('1', 'foo\n== section ==\nbar'))
         a = Argument('|\n==t==\nx')
-        self.assertEqual(
-            (a.name, a.value), ('1', '\n==t==\nx'))
+        ae((a.name, a.value), ('1', '\n==t==\nx'))
         # Following cases is not treated as a section headings
         a = Argument('|==1==\n')
-        self.assertEqual(
-            (a.name, a.value), ('', '=1==\n'))
+        ae((a.name, a.value), ('', '=1==\n'))
         # Todo: Prevents forming a template!
         # a = Argument('|\n==1==')
-        # self.assertEqual(
+        # ae(
         #     (a.name, a.value), ('1', '\n==1=='))
 
     def test_argument_name_not_external_link(self):
+        ae = self.assertEqual
         # MediaWiki parses template parameters before external links,
         # so it goes with the named parameter in both cases.
         a = Argument('|[http://example.com?foo=bar]')
-        self.assertEqual(
-            (a.name, a.value), ('[http://example.com?foo', 'bar]'))
+        ae((a.name, a.value), ('[http://example.com?foo', 'bar]'))
         a = Argument('|http://example.com?foo=bar')
-        self.assertEqual(
-            (a.name, a.value), ('http://example.com?foo', 'bar'))
+        ae((a.name, a.value), ('http://example.com?foo', 'bar'))
 
     def test_lists(self):
-        self.assertEqual(Argument('|list=*a\n*b').lists()[0].items, ['a', 'b'])
-        self.assertEqual(Argument('|lst= *a\n*b').lists()[0].items, ['a', 'b'])
-        self.assertEqual(Argument('|*a\n*b').lists()[0].items, ['a', 'b'])
+        ae = self.assertEqual
+        ae(Argument('|list=*a\n*b').lists()[0].items, ['a', 'b'])
+        ae(Argument('|lst= *a\n*b').lists()[0].items, ['a', 'b'])
+        ae(Argument('|*a\n*b').lists()[0].items, ['a', 'b'])
         # the space at the beginning of a positional argument should not be
         # ignored. (?)
-        self.assertEqual(Argument('| *a\n*b').lists()[0].items, ['b'])
+        ae(Argument('| *a\n*b').lists()[0].items, ['b'])
 
 
 if __name__ == '__main__':