+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ v0.25.0.dev0
 - Added a deleter for ``WikiLink.text`` property.
 - Deprecated: Setting ``WikiLink.text`` to None. Use ``del WikiLink.text`` instead.
 - Added a deleter for ``WikiLink.target`` property.
+- Added a deleter for ``Argument.value`` property.
 
 v0.24.4
 -------